requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
python-jose
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart
alembic